move dialog header to `<Dialog.Header>`
